[TD-225]refactor.
