+++ src/query/inc/qAst.h
@@ -74,9 +74,7 @@ typedef struct tExprNode {
   };
 } tExprNode;
 
-void tExprTreeTraverse(tExprNode *pExpr, SSkipList *pSkipList, SArray *result, SExprTraverseSupp *param);
-
-void tExprTreeCalcTraverse(tExprNode *pExprs, int32_t numOfRows, char *pOutput, void *param, int32_t order,
+void arithmeticTreeTraverse(tExprNode *pExprs, int32_t numOfRows, char *pOutput, void *param, int32_t order,
                                 char *(*cb)(void *, const char*, int32_t));
 
 tExprNode* exprTreeFromBinary(const void* data, size_t size);
@@ -86,6 +84,8 @@ void exprTreeToBinary(SBufferWriter* bw, tExprNode* pExprTree);
 
 void tExprNodeDestroy(tExprNode *pNode, void (*fp)(void *));
 void tExprTreeDestroy(tExprNode **pExprs, void (*fp)(void*));
+
+bool exprTreeApplayFilter(tExprNode *pExpr, const void *pItem, SExprTraverseSupp *param);
 
 #ifdef __cplusplus
 }
